pyproject.toml
[project]
name = "multimodal-fusion"
version = "0.1.0"
description = "Multi-image fusion and classification pipeline for the MURA dataset"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "kaggle>=1.7.4.5",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "timm>=1.0.25",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
]
